Inject specific service status inspector into health check controller
